WebApp/requirements.txt
flask==2.3.3
inference-sdk
Pillow
requests
flask-cors 